refactor: User 데이터를 response로 보낼 때 사용하는 UserDto 구현
비밀번호 같은 데이터를 response로 보내주지 않기 위해서 보낼 데이터만을 가지는 UserDto를 구현했습니다.
User 클래스에 toUserDto()를 구현해서 UserDto로 변경할 수 있도록 했습니다.
UserRepository에서는 User대신 UserDto를 리턴하도록 변경했습니다.
